+++ lib/lottie/objects/__init__.py
@@ -1,12 +1,16 @@
 """!
 Package with all the Lottie Python bindings
 """
-from . import animation, base, effects, enums, helpers, layers, shapes, assets, easing, text, bezier
+from . import (
+    animation, base, effects, enums, helpers, layers, shapes, assets, easing,
+    text, bezier, composition
+)
 from .animation import Animation
 from .layers import *
 from .shapes import *
 from .assets import Precomp
 from .bezier import Bezier
+from .composition import Composition
 
 __all__ = [
     "animation", "base", "effects", "enums", "helpers", "layers", "shapes", "assets"
@@ -15,5 +19,5 @@ __all__ = [
     "NullLayer", "TextLayer", "ShapeLayer", "ImageLayer", "PreCompLayer", "SolidLayer",
     "Rect", "Fill", "Trim", "Repeater", "GFill", "Stroke", "RoundedCorners", "Path",
     "TransformShape", "Group", "Star", "Ellipse", "Merge", "GStroke",
-    "Bezier", "Precomp",
+    "Bezier", "Precomp", "Composition",
 ]