+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased
+ * Support for all importers / exporters in lottie GUI
 
 ## 0.6.1 - 2020-04-25
  * Made scripts more user friendly